회원정보 수정 API 구현

+++ src/main/java/kr/adapterz/community/dto/UserInfoUpdateRequestDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class UserInfoUpdateRequestDto {
-    private Integer userId; // 세션 정보를 가져오게 되면 불필요
+    private Long userId; // 세션 정보를 가져오게 되면 불필요
     private String nickname;
     private String profileImage;
 }
